requirements.txt
kafka-python==2.3.0
numpy==2.4.2
pandas==3.0.1
pyarrow==23.0.1
python-dateutil==2.9.0
six==1.17.0
